+++ app/src/main/java/com/moringaschool/mywordfinder/models/WordsActivity.java
@@ -1,19 +1,18 @@
-package com.moringaschool.mywordfinder;
+package com.moringaschool.mywordfinder.models;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.DefaultItemAnimator;
-import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.moringaschool.mywordfinder.network.DataMuseApi;
+import com.moringaschool.mywordfinder.adapters.MyAdapter;
+import com.moringaschool.mywordfinder.R;
 
 import java.util.ArrayList;
 import java.util.List;
